requirements.txt
Flask==0.10.1
Werkzeug==0.9.1
Flask-Admin==1.0.8
Flask-Assets==0.10
Flask-Login==0.2.3
Flask-Mail==0.7.3
Flask-Principal==0.3.3
Flask-Restless==0.13.1
Flask-Script==0.5.3
Flask-Security==1.7.3
Flask-WTF==0.9.5
flask-s3==0.1.7
flask-classy==0.6.10
Flask-SQLAlchemy==1.0
itsdangerous==0.21
Jinja2==2.7
jsmin==2.0.3
cssmin==0.1.4
logilab-common==0.62.0
passlib==1.6.1
python-dateutil==2.1
python-mimeparse==0.1.4
simplejson==3.3.0
six==1.3.0
boto
webassets==0.10.1
WTForms==1.0.5
-e git+https://github.com/kvesteri/wtforms-json.git@877538171526b5272fa1e8229bec0831f14ce30f#egg=wtforms-json
